+++ lib/modules/vscode/transformVSCodeSettings.ts
@@ -2,7 +2,7 @@ import { readFile, writeFile } from "fs/promises";
 import ensureVSCodeSettings from "./ensureVSCodeSettings";
 import { join } from "path";
 
-export default async function transformVSCodeSettings (workdir: string, transform: (settings: any) => any) {
+export default async function transformVSCodeSettings (transform: (settings: any) => any, workdir: string = process.cwd()) {
     const settingsPath = join(workdir, '.vscode', 'settings.json');
 
     await ensureVSCodeSettings(workdir);