+++ CHANGELOG.md
@@ -6,7 +6,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-## [0.1.0] - 2024-04-07
+
+## [0.2.0] - 2024-04-12 (The Hidden files update)
+
+### Added
+- [2024-04-12] Created `hide`, `unhide`, `toggle` and `hidden` commands to manage the vscode files.exclude in a project
+
+
+### Changed
+- [2024-04-12] Changed `transformVSCodeSettings` parameter order, in order to allow the workdir to be optional defaulting to process.cwd()
+
+## [0.1.0] - 2024-04-07 (The Color Update)
 
 ### Added
 - [2024-04-07] Customization of bottom bar color is now possible via "peacock > color" setting